requirements.txt
# Core ML
torch==2.10.0+cu12 --extra-index-url https://download.pytorch.org/whl/cu121
torchvision==0.25.0+cu12 --extra-index-url https://download.pytorch.org/whl/cu121
timm==1.0.24
open_clip_torch==3.2.0

# Transformers & HuggingFace
transformers==5.1.0
tokenizers==0.22.2
huggingface_hub==1.4.1
safetensors==0.7.0
sentencepiece==0.2.1

# Data & ML utils
numpy==2.4.2
pandas==3.0.0
scikit-learn==1.8.0
scipy==1.17.0
pillow==12.1.1
h5py==3.16.0

# Training utilities
tqdm==4.67.3
tensorboard==2.20.0
kaggle==2.0.0

# General
requests==2.32.5
pyyaml==6.0.3
regex==2026.1.15
ftfy==6.3.1
packaging==26.0